pyproject.toml
[project]
name = "ai-factory"
version = "0.2.0"
description = "Unified AI operating system for managing the complete LLM lifecycle from training to deployment."
requires-python = ">=3.10"
dependencies = [
  "accelerate>=1.2.1",
  "datasets>=3.2.0",
  "fastapi>=0.115.6",
  "httpx>=0.28.1",
  "numpy>=2.1.3",
  "orjson>=3.10.12",
  "pandas>=2.2.3",
  "peft>=0.14.0",
  "pydantic>=2.10.4",
  "python-multipart>=0.0.20",
  "pyyaml>=6.0.2",
  "sympy>=1.13.3",
  "tokenizers>=0.21.0",
  "torch>=2.5.1",
  "transformers>=4.47.1",
  "trl>=0.12.0",
  "tqdm>=4.67.1",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
train-cuda = [
  "bitsandbytes>=0.45.0",
]
macos = [
  "mlx>=0.28.0; platform_system == 'Darwin' and platform_machine == 'arm64'",
  "mlx-lm>=0.20.6; platform_system == 'Darwin' and platform_machine == 'arm64'",
  "sentencepiece>=0.2.0",
]
dev = [
  "pre-commit>=4.0.1",
  "mypy>=1.15.0",
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.3",
  "pytest-cov>=6.0.0",
  "ruff>=0.15.7",
  "types-PyYAML>=6.0.12",
  "types-requests>=2.32.0",
  "bandit>=1.7.0",
  "safety>=3.0.0",
]

[project.scripts]
ai-factory = "ai_factory.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["ai_factory*", "training*", "inference*", "evaluation*", "data*"]
exclude = ["tests", "tests.*"]

[build-system]
requires = ["setuptools>=75.6.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
extend-ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"inference/app/routers/*.py" = ["B008"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_return_any = false
warn_no_return = false
disallow_untyped_defs = true
exclude = ["^tests/"]

[[tool.mypy.overrides]]
module = [
    "data.*",
    "evaluation.*",
    "training.*",
    "inference.*",
    "ai_factory.interfaces.*",
    "ai_factory.cli",
    "tests.*"
]
ignore_errors = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-ra"
